Changes shooter talon to speed control mode, trying to find the correct canid for the intake

+++ 2017CompetitionBot/src/Commands/LiftCmd.cpp
@@ -1,6 +1,6 @@
 #include "LiftCmd.h"
 
-LiftCmd::LiftCmd(int speed) {
+LiftCmd::LiftCmd(double speed) {
 	// Use Requires() here to declare subsystem dependencies
 	// eg. Requires(Robot::chassis.get());
 	Requires(liftSub.get());
